services/chronicle-service/src/requirements.txt
fastapi
uvicorn
psycopg2-binary
structlog
redis
httpx
